Create Professor Grading System
Environment
Our project was designed with the following environment established: the script, files, and directory used within code is located within a directory titled project that is located within the personal directory of the professor who we identify as the user. Within the project folder there’s a script that’ll execute our code titled “grading.sh” , a file titled “gradebook” that will be there once the professor executed code for the first time to create a new gradebook, a file titled “oldgradebook” that is a reference to professor's prior gradebook before executing grade.sh code (the oldgradebook is used within one of the functionalities), and a directory titled “students” that contains further directories of the students. The student’s directory is meant to contain the directories of all the students within the course, with the directories of these students holding their own homework submissions. Each student directory can contain zero to many homework submissions that the script will be able to automatically grade.

Script Functionalities
Upon execution of the script, the user is greeted with a message to select the functionality of their choice, via input on the command line. All user inputs via command line are secured with a while loop that will continue to iterate accessing for user input until the user writes an acceptable input.
Grade	
The grade functionality is meant to act as our automatic grader. Upon selecting this functionality, the user will be asked if they would like to create a new gradebook or update an existing gradebook. If the user selects new, the script will create a new file titled “Gradebook” that will write a column titled “students”, which consists of all the names of the students identified within the student directory. Following this, the user is asked what submission they would like to grade for all their students. Upon choosing the submission, the script will create a temporary file that consists of a column that is titled of the submission. After selecting the file, the user is asked how many homework problems they would like to grade within the homework submission. Following this, the program will then ask for the solutions of these problems and will store these within another temporary file. Once the problems and their solutions are determined and written within temporary files, the program will begin to grade the students submissions using a for loop, with the added use of a while loop to read through the lines of problems within the submission. As the student code is tested using the eval command, the student output and the professor output is compared using the diff command as a basis of scoring. If there’s no difference, then the student receives a full two points. If there’s differences that exceed over twenty lines or if the differences are equal to that of the number of lines of output, then the student is given a zero. If neither of those conditions are meant, then the student is given partial credit; with the student earning two points and losing a tenth of a point for every wrong line. Once a problem is graded, the script will continue to iterate through the remainder of problems, keeping track of the overall grade on the assignment. After all the submissions have been graded for a student, then the total grade is saved within a temporary file and the script moves on to the next student repeating the same process. Once all the students have been graded and written within the temporary file, we  use the paste command to insert the column in the temporary file into the gradebook file, using the comma as a delimiter. After all these steps are completed, this functionality will have concluded with all the students' submission for the user defined assignment graded and written in the gradebook.
Change 
The change functionality is meant to change the grade of any singular assignment for any student within the course, in the case that the professor wants to change the grade of a student from a prior assignment. This functionality begins by outputting the gradebook as it is now and asks the user to select the assignment in specific they would like to change the grade of for a student. Following this, the code then asks what is the name of the student whose grade you’d like to change. Both the student and assignment selection is protected using a while loop,  only allowing for appropriate values, otherwise the code won’t continue. Following the users input of a new grade, using sed, the script will change that specified grade in the gradebook and output the new modified gradebook with the changed grade.
Update
The update functionality is meant to update the new gradebook the professor/user is using in the case that there is an old gradebook the professor/user would like to extract grade from. This functionality begins by asking the user what is the name of the old gradebook they would like to pull a column of grades from. Once the old gradebook is determined by the script, the script then asks the user to input the column of assignments and their corresponding grade they would like to paste into the new gradebook. Once selected, the new updated gradebook will be output with the addition of the previous course assignment and its grades.